Working requests page. - fixed queries - fixed sorting and filtering
